Add styling to featured items
This is pretty ugly right now, but it's a good starting point. This puts
the featured items in bordered boxes and makes the whole thing
clickable. The hack to make the whole thing clickable places and empty
link inside the div that contains a span element which is styled to fill
the entire div.